backend/pyproject.toml
[project]
name = "counselai-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "qdrant-client>=1.13",
    "pymupdf>=1.25",
    "python-multipart>=0.0.18",
]
